Refactor CanvasDrawerPanel to enhance state management and UI integration. Move grouped shapes logic from Tldraw to CanvasDrawerPanel, simplifying component structure and improving maintainability.
