+++ apps/roam/src/components/canvas/CanvasDrawer.tsx
@@ -3,6 +3,7 @@ import {
   Button,
   Card,
   Collapse,
+  Icon,
   Menu,
   MenuItem,
   NonIdealState,
@@ -14,7 +15,9 @@ import {
   Tag,
   Tooltip,
 } from "@blueprintjs/core";
+import { Editor } from "tldraw";
 import getPageTitleByPageUid from "roamjs-components/queries/getPageTitleByPageUid";
+import getCurrentPageUid from "roamjs-components/dom/getCurrentPageUid";
 import getDiscourseNodes from "~/utils/getDiscourseNodes";
 import { DiscourseNodeShape } from "./DiscourseNodeUtil";
 import { formatHexColor } from "~/components/settings/DiscourseNodeCanvasSettings";
@@ -370,5 +373,105 @@ export const CanvasDrawerContent = ({ groupedShapes, pageUid }: Props) => {
         </Card>
       )}
     </div>
+  );
+};
+
+type CanvasDrawerPanelProps = {
+  editor: Editor;
+  isOpen: boolean;
+  onToggle: () => void;
+};
+
+export const CanvasDrawerPanel = ({
+  editor,
+  isOpen,
+  onToggle,
+}: CanvasDrawerPanelProps) => {
+  const pageUid = getCurrentPageUid();
+  const [groupedShapes, setGroupedShapes] = useState<GroupedShapes>({});
+
+  useEffect(() => {
+    const updateGroupedShapes = () => {
+      const allRecords = editor.store.allRecords();
+      const shapes = allRecords.filter((record) => {
+        if (record.typeName !== "shape") return false;
+        const shape = record as DiscourseNodeShape;
+        return !!shape.props?.uid;
+      }) as DiscourseNodeShape[];
+
+      const grouped = shapes.reduce((acc: GroupedShapes, shape) => {
+        const uid = shape.props.uid;
+        if (!acc[uid]) acc[uid] = [];
+        acc[uid].push(shape);
+        return acc;
+      }, {});
+
+      setGroupedShapes(grouped);
+    };
+
+    updateGroupedShapes();
+
+    const unsubscribe = editor.store.listen(() => {
+      updateGroupedShapes();
+    });
+
+    return () => {
+      unsubscribe();
+    };
+  }, [editor.store]);
+
+  return (
+    <>
+      <div
+        className="pointer-events-auto absolute top-11 m-2 rounded-lg"
+        style={{
+          zIndex: 250,
+          // copying tldraw var(--shadow-2)
+          boxShadow:
+            "0px 0px 2px hsl(0, 0%, 0%, 16%), 0px 2px 3px hsl(0, 0%, 0%, 24%), 0px 2px 6px hsl(0, 0%, 0%, 0.1), inset 0px 0px 0px 1px hsl(0, 0%, 100%)",
+          backgroundColor: "white",
+        }}
+      >
+        <Button
+          icon={<Icon icon="add-column-left" />}
+          onClick={onToggle}
+          minimal
+          title="Toggle Canvas Drawer"
+        />
+      </div>
+      {isOpen && (
+        <div
+          className="pointer-events-auto absolute left-0 flex w-80 flex-col bg-white shadow-lg"
+          style={{
+            top: "40px",
+            height: "calc(100% - 40px)",
+            zIndex: 250,
+            boxShadow:
+              "0px 0px 2px hsl(0, 0%, 0%, 16%), 0px 2px 3px hsl(0, 0%, 0%, 24%), 0px 2px 6px hsl(0, 0%, 0%, 0.1), inset 0px 0px 0px 1px hsl(0, 0%, 100%)",
+          }}
+        >
+          <div className="flex max-h-10 flex-shrink-0 items-center justify-between overflow-hidden border-b border-gray-300 bg-white px-3">
+            <h2 className="m-0 text-sm font-semibold leading-tight">
+              Canvas Drawer
+            </h2>
+            <div className="flex-shrink-0">
+              <Button
+                icon={<Icon icon="cross" />}
+                onClick={onToggle}
+                minimal
+                small
+                style={{ minHeight: 0, height: "24px", padding: "4px" }}
+              />
+            </div>
+          </div>
+          <div className="flex min-h-0 flex-1 flex-col overflow-hidden p-4">
+            <CanvasDrawerContent
+              groupedShapes={groupedShapes}
+              pageUid={pageUid}
+            />
+          </div>
+        </div>
+      )}
+    </>
   );
 };
